pyproject.toml
[project]
name = "osrsflipfinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "discord-py>=2.4.0",
    "fuzzywuzzy>=0.18.0",
    "mypy>=1.15.0",
    "osrsreboxed>=2.3.33",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "texttable>=1.7.0",
    "types-requests>=2.32.0.20241016",
    "uvicorn>=0.34.0",
]
